backend/requirements.txt
# Django и основные зависимости
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# База данных
mysqlclient==2.2.0

# Авторизация
djangorestframework-simplejwt==5.3.1

# Фоновые задачи
celery==5.3.4
redis==5.0.1

# Интеграции
python-telegram-bot==20.7
requests==2.31.0
openai==1.12.0
# Instagram Basic Display API использует requests напрямую

# Утилиты
python-dotenv==1.0.0
Pillow==10.2.0

# Безопасность
cryptography==42.0.2

# Production Server
gunicorn==21.2.0
